feat: internal of pacckage
