--- app/domain/model.go
@@ -1,19 +1,0 @@
-package domain
-
-import "time"
-
-type Person struct {
-	ID        int64
-	FirstName string
-	LastName  string
-	SurName   string
-	Login     string
-}
-
-type Course struct {
-	ID        int64
-	Title     string
-	Status    string 
-	StartDate time.Time
-	EndDate   time.Time
-}
