+++ internal/domain/course.go
@@ -1,0 +1,11 @@
+package domain
+
+import "time"
+
+type Course struct {
+	ID        int64
+	Title     string
+	Status    string
+	StartDate time.Time
+	EndDate   time.Time
+}
