+++ internal/domain/person.go
@@ -1,0 +1,9 @@
+package domain
+
+type Person struct {
+	ID        int64
+	FirstName string
+	LastName  string
+	SurName   string
+	Login     string
+}
