+++ internal/usecases/user.go
@@ -4,8 +4,8 @@ import (
 	"context"
 	"fmt"
 
-	"github.com/lozovoya/GolangUnitedSchool/app/domain"
-	"github.com/lozovoya/GolangUnitedSchool/app/repository"
+	"github.com/lozovoya/GolangUnitedSchool/internal/domain"
+	"github.com/lozovoya/GolangUnitedSchool/internal/repository"
 )
 
 type UserSt struct {
